updated message component , adding provisions for approve/disapproving photos to admins, fixed issue with datetime stored in database and confirmation modal

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,8 @@ export class MessagesComponent implements OnInit {
   pagination: Pagination;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService,
+     private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -31,10 +33,13 @@ export class MessagesComponent implements OnInit {
     this.loading = false;
   }
 
-
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m=>m.id === id), 1)
+    this.confirmService.confirm('Confirm delete message').subscribe(result => {
+      if(result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m=>m.id === id), 1)
+        })
+      }
     })
   }
